Removed c++11 stuff

+++ image_filter/pthread/thresmain.cc
@@ -90,7 +90,7 @@ int main (int argc, char ** argv) {
 
   clock_gettime(CLOCK_REALTIME, &etime);
 
-  const int delta_time = (etime.tv_sec  - stime.tv_sec) + 1e-9*(etime.tv_nsec  - stime.tv_nsec);
+  const double delta_time = (etime.tv_sec  - stime.tv_sec) + 1e-9*(etime.tv_nsec  - stime.tv_nsec);
   cout << "Filtering took: "<< delta_time <<" secs" << endl;
   
   /* write result */
